zad4 done

+++ lab4/Catcher.c
@@ -2,19 +2,40 @@
 #include<stdlib.h>
 #include<unistd.h>
 #include<signal.h>
+#include<string.h>
 
 #define TRUE 1
 
 int caught;
+char mode[256];
 
 void sigUsr1Handle(int sig, siginfo_t* info, void* ucontext){
     caught++;
-    kill(info->si_pid, SIGUSR1);
+    if(strcmp(mode, "kill") == 0){
+        kill(info->si_pid, SIGUSR1);
+    }
+    else if(strcmp(mode, "sigqueue") == 0){
+        union sigval val;
+        val.sival_int = caught;
+        sigqueue(info->si_pid, SIGUSR1, val);
+    }
+    else{
+        kill(info->si_pid, SIGRTMIN+1);
+    }
 }
 
 void sigUsr2Handle(int sig, siginfo_t* info, void* ucontext){
     printf("%s %d %s\n","catcher got", caught, "signals");
-    kill(info->si_pid, SIGUSR2);
+    if(strcmp(mode, "kill") == 0){
+        kill(info->si_pid, SIGUSR2);
+    }
+    else if(strcmp(mode, "sigqueue") == 0){
+        union sigval val;
+        sigqueue(info->si_pid, SIGUSR2, val);
+    }
+    else{
+        kill(info->si_pid, SIGRTMIN+2);
+    }
     exit(0);
 }
 
@@ -27,17 +48,29 @@ void setUpSignalHandling(){
     action1.sa_sigaction = sigUsr2Handle;
     sigemptyset(&action.sa_mask);
     sigemptyset(&action1.sa_mask);
-    sigaction(SIGUSR1, &action, NULL);
-    sigaction(SIGUSR2, &action1, NULL);
+    if(strcmp(mode, "SIGRT")!=0){
+        sigaction(SIGUSR1, &action, NULL);
+        sigaction(SIGUSR2, &action1, NULL);
+    }
+    else{
+        sigaction(SIGRTMIN+1, &action, NULL);
+        sigaction(SIGRTMIN+2, &action1, NULL);
+    }
     sigset_t set;
     sigfillset(&set);
     sigdelset(&set, SIGUSR1);
     sigdelset(&set, SIGUSR2);
+    sigdelset(&set, SIGINT);
+    if(strcmp(mode, "SIGRT") == 0){
+        sigdelset(&set, SIGRTMIN+1);
+        sigdelset(&set, SIGRTMIN+2);
+    }
     sigprocmask(SIG_SETMASK, &set, NULL);
 }
 
 
-int main(){
+int main(int argc, char* argv[]){
+    strcat(mode, argv[1]);
     caught = 0;
     printf("%s %d\n","my pid", getpid());
     setUpSignalHandling();
